% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembleToDf.R
\name{assembleToDf}
\alias{assembleToDf}
\title{Gives nice dataframe}
\usage{
assembleToDf(tableList, df, id, tsne_results)
}
\arguments{
\item{tableList}{list of final tableing to use}

\item{df}{original dataset used}

\item{id}{What is the Id column in the original dataset. Should be the same}

\item{tsne_results}{Tsne data(X, Y, cluster) from plot_clustering object
id column in the final tableing list.}
}
\description{
Gives nice dataframe of personId, their cluster, their table, their tsne
coordinates, and their original responses to the questions.
}
\keyword{Arrange}
\keyword{and}
\keyword{clusters}
\keyword{for}
\keyword{including}
\keyword{nice}
\keyword{numbers}
\keyword{table}
