% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPamModel.R
\name{buildPamModel}
\alias{buildPamModel}
\title{Build a PAM model}
\usage{
buildPamModel(dissimilarities, silWidths = NULL, k = NULL)
}
\arguments{
\item{dissimilarities}{gower dissimilarity to use}

\item{silWidths}{Allows for use of a vector of silWidths. Inherited from
silWidth$widths. You must specify a k or vector of widths.}

\item{k}{Number of clusters to find.  If silWidths is present, k will become
which.max(silWidths). You must specify a k or vector of widths.}
}
\description{
This function allows user to build a PAM model
}
\keyword{Dissimilarity}
\keyword{Gower,}
\keyword{PAM,}
