% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterSummary.R
\name{clusterSummary}
\alias{clusterSummary}
\title{Compute Cluster summary from PAM}
\usage{
clusterSummary(df, id = NULL, pam_model)
}
\arguments{
\item{df}{Data frame from which dissimilarities were calculated}

\item{id}{If there is an id column, what is it?}

\item{pam_model}{model built previously using pamModel()}
}
\description{
This function allows user to compute cluster summaries for a PAM model
}
\keyword{Cluster}
\keyword{Results}
\keyword{Summary,}
