% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeGowerDiss.R
\name{computeGowerDiss}
\alias{computeGowerDiss}
\title{Compute Gower Dissimilarity}
\usage{
computeGowerDiss(df, id = NULL)
}
\arguments{
\item{df}{Data frame to use in dissimilarity computation}

\item{id}{If there is an id column, what is it?}
}
\description{
This function allows user to compute gower dissimilarity for a dataframe
}
\keyword{Dissimilarity}
