% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTsneAnimation.R
\name{createTsneAnimation}
\alias{createTsneAnimation}
\title{Animation of plotting PAM cluster results in 2D space using t-SNE}
\usage{
createTsneAnimation(df, gowerDiss, id, pam_fit, iterations)
}
\arguments{
\item{df}{Data frame from which dissimilarities were calculated}

\item{gowerDiss}{gower dissimilarities to use}

\item{id}{ids from original data}

\item{pam_fit}{PAM object from building a PAM clustering model}

\item{iterations}{Number of iterations for tsne}
}
\description{
This function allows user to plot cluster results from a PAM object using
t-distributed stochastic neighborhood embedding.
}
\keyword{clustering,}
\keyword{dimensionality}
\keyword{plot}
\keyword{reduction}
\keyword{t-SNE,}
