% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTsneSlides.R
\name{createTsneSlides}
\alias{createTsneSlides}
\title{Create slides of tsne process for use in animations}
\usage{
createTsneSlides(df, gowerDiss, id, pam_fit, iterations)
}
\arguments{
\item{df}{Data frame from which dissimilarities were calculated}

\item{gowerDiss}{gower dissimilarities to use}

\item{id}{ids from original data}

\item{pam_fit}{PAM object from building a PAM clustering model}

\item{iterations}{Number of iterations for tsne}
}
\description{
This function allows user to plot cluster results from a PAM object using
t-distributed stochastic neighborhood embedding. Several slides are created
based on the number of iterations specified. E.g. If you specify 1000
iterations, the slides created will be tsne results from 1, 10, 20, 30, 40,
and so on up to 1000 iterations.
}
\keyword{clustering,}
\keyword{dimensionality}
\keyword{plot}
\keyword{reduction}
\keyword{t-SNE,}
