% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-find.R
\name{findLeastSimilar}
\alias{findLeastSimilar}
\title{Compute Least Similar Observations}
\usage{
findLeastSimilar(diss, df)
}
\arguments{
\item{diss}{dissimilarity object}

\item{df}{dataframe from which the dissimilarity object was calculated}
}
\description{
This function allows user to find the least similar observations from the
dissimilarity matrix
}
\keyword{Dissimilarity,}
\keyword{Find}
\keyword{Least}
