% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-find.R
\name{findMostSimilar}
\alias{findMostSimilar}
\title{Compute Most Similar Observations}
\usage{
findMostSimilar(diss, df)
}
\arguments{
\item{diss}{dissimilarity object}

\item{df}{dataframe from which the dissimilarity object was calculated}
}
\description{
This function allows user to find most similar observations from the
dissimilarity matrix
}
\keyword{Dissimilarity,}
\keyword{Find}
\keyword{Most}
