% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatherTables.R
\name{gatherTables}
\alias{gatherTables}
\title{Put as many people into specified table size}
\usage{
gatherTables(df, id, pam_model, tableSize)
}
\arguments{
\item{df}{original dataframe from which the clustering was done}

\item{id}{Id column from dataset}

\item{pam_model}{PAM object from building a pam model}

\item{tableSize}{How many people per table are there going to be?}
}
\value{
A list of full tables and a list of tables that are not full,
referred to as leftOvers.
}
\description{
This function is the first iteration of tableing
}
\keyword{Tabling}
