% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfMetrics.R
\name{perfMetrics}
\alias{perfMetrics}
\title{Performance Metrics of the Tabling}
\usage{
perfMetrics(merged_data)
}
\arguments{
\item{merged_data}{data set with cluster, table, and tsne data
from assembleToFinalDf}
}
\description{
Gives performance metrics of tabling and compiles a finished dataset with
the performance metrics given at the grain level.
}
\keyword{Performance}
\keyword{metrics}
