% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTables.R
\name{plotTables}
\alias{plotTables}
\title{Plot Results of Tabling}
\usage{
plotTables(merged_data)
}
\arguments{
\item{merged_data}{Final dataframe that includes table, cluster, and tsne
assignments}
}
\description{
Plot the results of the finished product/dataframe from tableing
}
\keyword{Arrange}
\keyword{and}
\keyword{clusters}
\keyword{for}
\keyword{including}
\keyword{nice}
\keyword{numbers}
\keyword{table}
